# 9337 zfs get all is slow due to uncached metadata