Reviewed by: Prakash Surya <prakash.surya@delphix.com>
Reviewed by: George Wilson <george.wilson@delphix.com>

This project's goal is to make read-heavy channel programs and zfs(1m)
administrative commands faster by caching all the metadata that they will
need in the dbuf layer. This will prevent the data from being evicted, so
that any future call to i.e. zfs get all won't have to go to disk (very
much). There are two parts:

1) The dbuf_metadata_cache. We identify what to put into the cache based on
the object type of each dbuf.  
2) Caching objset properties os{version,normalization,utf8only,casesensitivity} 
in the objset_t. The reason these needed to be cached is that although they are 
queried frequently, they aren't stored in a dbuf type which we can easily recognize 
and cache in the dbuf layer; instead, we have to explicitly store them. There's 
alreadyexisting infrastructure for maintaining cached properties in the objset
setup code, so I simply used that.

Testing performance
Setup:
Disabled kmem_flags
Tuned dbuf_cache_max_bytes very low (128K)
Tuned zfs_arc_max very low (64M)
Created test pool with 400 filesystems, and 100 snapshots per filesystem.
Later on in testing, added 600 more filesystems (with no snapshots) to make
sure scaling didn't look different between snapshots and filesystems.

Results:

```
| Test                           | Time (trunk / diff) | I/Os (trunk / diff) |
+--------------------------------+---------------------+---------------------+
| zpool import                   |     0:05 / 0:06     |    12.9k / 12.9k    |
| zfs get all (uncached)         |     1:36 / 0:53     |    16.7k / 5.7k     |
| zfs get all (cached)           |     1:36 / 0:51     |    16.0k / 6.0k     |
```
Most cases are faster, and we do a reasonable amount less I/O in every case (except for zpool import where we weren't expecting an improvement). Great :)

Upstream bug: DLPX-28586

Reviewed at: http://reviews.delphix.com/r/34845/