@ahrens 2018-03-23
@tcaputi To determine the cache size, you could look at `dbuf_caches[DB_DBUF_METADATA_CACHE].size`.  While this is trivial with mdb (assuming you have root access), I agree it would be nice to add a kstat for this.  IIRC, it seemed like kind of a pain since the kstats seem to be specific to each .c file, and there isn't already a kstat set up in dbuf.c.  But that doesn't mean we shouldn't do it at some point :)