@richardelling 2018-03-25
In ZoL there is a dbuf_stats kstat that is the appropriate place for these. It will be natural to extend for the ZoL port.